+++ pdfalyzer/config.py
@@ -12,7 +12,7 @@ from yaralyzer.util.constants import MAX_FILENAME_LENGTH
 from yaralyzer.util.logging import log
 
 from pdfalyzer.detection.yaralyzer_helper import YARA_RULES_FILES
-from pdfalyzer.output.theme import PDFALYZER_THEME_DICT, THEME_COLORS_FOR_SHOW_ONLY_DICT
+from pdfalyzer.output.theme import COMPLETE_THEME_DICT
 from pdfalyzer.util.constants import PDFALYZE, PDFALYZER_UPPER
 from pdfalyzer.util.helpers.filesystem_helper import DEFAULT_PDF_PARSER_PATH, PDF_PARSER_PATH_ENV_VAR
 from pdfalyzer.util.output_section import ALL_STREAMS
@@ -33,7 +33,7 @@ class PdfalyzerConfig(YaralyzerConfig):
 
     # Override the class vars of same name in YaralyzerConfig
     ENV_VAR_PREFIX = PDFALYZER_UPPER
-    COLOR_THEME = {**PDFALYZER_THEME_DICT, **THEME_COLORS_FOR_SHOW_ONLY_DICT}
+    COLOR_THEME = COMPLETE_THEME_DICT
     ONLY_CLI_ARGS = YaralyzerConfig.ONLY_CLI_ARGS + ['extract_binary_streams']
 
     pdf_parser_path: Path | None = None