+++ scripts/use_local_yaralyzer.sh
@@ -16,5 +16,10 @@ if [[ ! -z $local_yaralyzer_branch ]]; then
 fi
 
 update_pyproject_toml $YARALYZER "$LOCAL_YARALYZER_REQ"
-poetry lock
+
+if ! poetry lock; then
+    echo_warning "Poetry lock failed, trying with --no-cache option..."
+    poetry lock --no-cache
+fi
+
 poetry install --all-extras