+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # NEXT RELEASE
+* `/AA:JavaScript` nodes get the red blinking alert treatment
+* Refactor overall theme / color / style situation for PDF nodes
+* Extra debug output for `--show-colors --debug` option combinattion
 
 ### 1.19.2
 * Use `pdf-parser.py` to verify object IDs instead of checking all IDs from 0 to `num_nodes`